Rebased on top of #14705
don't return VMI after patching, as this could corrupt the
informer cache as we are passing it in as is.
don't check for nil annotation map as getting k,v from it
won't npe as I thought.

update the runStrategy logic for 'Always' to actuall stop
the VM if we detect it was a decentralized migration. This
is to prevent VMs that are migrated from restarting.

+++ pkg/virt-operator/resource/apply/core.go
@@ -808,7 +808,7 @@ func (r *Reconciler) updateSynchronizationAddress() (err error) {
 		if err != nil {
 			return err
 		}
-		port = p
+		port = int32(p)
 	}
 	r.kv.Status.SynchronizationAddress = pointer.P(fmt.Sprintf("%s:%d", ip, port))
 	return nil
